requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymongo[srv]==3.11
python-dotenv==1.1.0
pandas==2.2.3